moved addmodel to flint.predict
