+++ flint/calibrate/aocalibrate.py
@@ -9,7 +9,6 @@ from typing import (
     Any,
     Collection,
     Iterable,
-    Literal,
     NamedTuple,
 )
 
@@ -30,25 +29,11 @@ from flint.bptools.smoother import (
 )
 from flint.exceptions import PhaseOutlierFitError
 from flint.logging import logger
-from flint.ms import consistent_ms, get_beam_from_ms, remove_columns_from_ms
+from flint.ms import consistent_ms, get_beam_from_ms
 from flint.naming import get_aocalibrate_output_path
 from flint.options import MS, BaseOptions
 from flint.sclient import run_singularity_command
 from flint.utils import create_directory
-
-
-class AddModelOptions(BaseOptions):
-    """Container for options into the ``addmodel`` program packaged
-    with ``aocalibrate``"""
-
-    model_path: Path
-    """Path to the sky-model that will be inserted"""
-    ms_path: Path
-    """Path to the measurement set that will be interacted with"""
-    mode: Literal["a", "s", "c", "v"]
-    """The mode ``addmodel`` will be operating under, where where a=add model to visibilities (default), s=subtract model from visibilities, c=copy model to visibilities, z=zero visibilities"""
-    datacolumn: str
-    """The column that will be operated against"""
 
 
 class CalibrateOptions(BaseOptions):
@@ -1082,53 +1067,6 @@ def flag_aosolutions(
     )
 
     return flagged_aosolutions
-
-
-def add_model_options_to_command(add_model_options: AddModelOptions) -> str:
-    """Generate the command to execute ``addmodel``
-
-    Args:
-        add_model_options (AddModelOptions): The collection of supported options used to generate the command
-
-    Returns:
-        str: The generated addmodel command
-    """
-    logger.info("Generating addmodel command")
-    command = f"addmodel -datacolumn {add_model_options.datacolumn} -m {add_model_options.mode} "
-    command += f"{add_model_options.model_path!s} {add_model_options.ms_path!s}"
-
-    return command
-
-
-def add_model(
-    add_model_options: AddModelOptions, container: Path, remove_datacolumn: bool = False
-) -> AddModelOptions:
-    """Use the ``addmodel`` program to predict the sky-model visibilities
-    from a compatible source list (e.g. ``wsclean -save-source-list``)
-
-    Args:
-        add_model_options (AddModelOptions): The set of supported options to be supplied to ``addmodel``
-        container (Path): The calibrate container that contains the ``addmodel`` program
-        remove_datacolumn (bool, optional): Whether to first removed the ``datacolumn`` specified in ``add_model_options`` before predicting. If False it should be overwritten. Defaults to False.
-
-    Returns:
-        AddModelOptions: The options used to run ``addmodel`` (same as input)
-    """
-    if remove_datacolumn:
-        remove_columns_from_ms(
-            ms=add_model_options.ms_path, columns_to_remove=add_model_options.datacolumn
-        )
-    add_model_command = add_model_options_to_command(
-        add_model_options=add_model_options
-    )
-
-    run_singularity_command(
-        image=container,
-        command=add_model_command,
-        bind_dirs=[add_model_options.ms_path, add_model_options.model_path],
-    )
-
-    return add_model_options
 
 
 def get_parser() -> ArgumentParser:
